Add getters to messages
